requirements.txt
alembic==1.11.1
Babel==2.12.1
blinker==1.6.2
click==8.1.3
Flask==2.3.2
Flask-BabelEx==0.9.4
Flask-JWT-Extended==4.5.2
Flask-Login==0.6.2
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
gevent==22.10.2
greenlet==2.0.2
h11==0.14.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
passlib==1.7.4
pydantic==1.10.8
PyJWT==2.7.0
python-dotenv==1.0.0
speaklater==1.3
SQLAlchemy==2.0.15
typing-extensions==4.6.2
unicorn==2.0.1.post1
uvicorn==0.22.0
Werkzeug==2.3.4
WTForms==3.0.1
